refactor(cockpit): extract markdown renderer
